CORE-8792 Move delete group confirmation to presenter
